+++ RawgMockApp/Model/SearchModel/SearchModel.swift
@@ -18,7 +18,7 @@ struct result:Codable,Identifiable {
     var name:String?
     var released:String?
     var image:String?
-    var rating:Float?
+    var rating:Double?
     var rank:Int?
     
     enum CodingKeys: String, CodingKey {
